client/requirements.txt
mido==1.2.10
psutil==5.8.0
python-rtmidi==1.4.9
redis==3.5.3
